requirements.txt
azure-ai-documentintelligence==1.0.0b1
azure-core>=1.29.4
openai>=1.3.0
python-dotenv>=1.0.0
requests>=2.31.0     
numpy>=1.24.0         
